organize perk patches into namespaces and directories
export and seal all patch classes that will have no descendants
