+++ src/CommunityPatch/CommunityPatch.csproj
@@ -60,6 +60,6 @@
         <PackageReference Include="Tomlyn" Version="0.1.2" />
     </ItemGroup>
     <ItemGroup>
-        <Folder Include="Patches\Perks\Leadership" />
+        <Folder Include="Patches\Perks\Intelligence" />
     </ItemGroup>
 </Project>
