+++ src/CommunityPatch/Patches/Perks/Control/Bow/MountedArcherPatch.cs
@@ -11,16 +11,16 @@ using TaleWorlds.Engine;
 using TaleWorlds.Library;
 using TaleWorlds.MountAndBlade;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Control.Bow {
 
-  public class RidingBowExpertPatch : AgentWeaponEquippedPatch<RidingBowExpertPatch> {
+  public sealed class MountedArcherPatch : AgentWeaponEquippedPatch<MountedArcherPatch> {
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(RidingBowExpertPatch).GetMethod(nameof(Prefix), BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo PatchMethodInfo = typeof(MountedArcherPatch).GetMethod(nameof(Prefix), BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
 
-    private static PerkObject _bowExpert;
+    private static PerkObject _mountedArcher;
 
     public override void Reset() {
-      _bowExpert = PerkObject.FindFirst(perk => perk.Name.GetID() == "cKTeea27");
+      _mountedArcher = PerkObject.FindFirst(perk => perk.Name.GetID() == "eU0uANvZ");
       base.Reset();
     }
 
@@ -39,9 +39,10 @@ namespace CommunityPatch.Patches {
     [MethodImpl(MethodImplOptions.NoInlining)]
     private static void Prefix(ItemMenuVM __instance, MBBindingList<ItemFlagVM> list, ref WeaponComponentData weapon) {
       var character = (BasicCharacterObject) ItemMenuVmCharacterField.GetValue(__instance);
-      if (weapon.ItemUsage == "long_bow") // Make sure we're always using the correct value, in case this overwrites some shared WeaponComponentData
+      // Make sure we're always using the correct value, in case this overwrites some shared WeaponComponentData
+      if (weapon.ItemUsage == "long_bow")
         WeaponComponentDataItemUsageMethod
-          .Invoke(weapon, new[] {HeroHasPerk(character, _bowExpert) ? "bow" : weapon.ItemUsage});
+          .Invoke(weapon, new object[] {HeroHasPerk(character, _mountedArcher) ? "bow" : weapon.ItemUsage});
     }
 
     protected override bool AppliesToVersion(Game game)
@@ -65,7 +66,7 @@ namespace CommunityPatch.Patches {
       for (var i = 0; i < weaponStatsData.Length; i++) {
         var weapon = weaponStatsData[i];
         if (weapon.ItemUsageIndex != MBItem.GetItemUsageIndex("long_bow")
-          || !HeroHasPerk(__instance.Character, _bowExpert))
+          || !HeroHasPerk(__instance.Character, _mountedArcher))
           continue;
 
         var updatedWeapon = weapon;
