+++ src/CommunityPatch/Patches/Perks/Endurance/Athletics/PeakFormPatch.cs
@@ -1,24 +1,24 @@
-using System;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Runtime.CompilerServices;
-using TaleWorlds.CampaignSystem;
-using TaleWorlds.Core;
-using TaleWorlds.Localization;
 using HarmonyLib;
+using Helpers;
+using TaleWorlds.CampaignSystem;
+using TaleWorlds.CampaignSystem.SandBox.GameComponents;
+using TaleWorlds.Core;
+using TaleWorlds.Library;
+using TaleWorlds.Localization;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Endurance.Athletics {
 
-  sealed class StewardAssessorPatch : PatchBase<StewardAssessorPatch> {
+  public sealed class PeakFormPatch : PatchBase<PeakFormPatch> {
 
     public override bool Applied { get; protected set; }
 
-    private static readonly MethodInfo TargetMethodInfo =
-      Type.GetType("TaleWorlds.CampaignSystem.SandBox.GameComponents.DefaultSettlementTaxModel, TaleWorlds.CampaignSystem")?
-        .GetMethod("CalculateDailyTaxInternal", BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo TargetMethodInfo = typeof(DefaultCharacterStatsModel).GetMethod(nameof(DefaultCharacterStatsModel.MaxHitpoints), BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(StewardAssessorPatch).GetMethod(nameof(Postfix), BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo PatchMethodInfo = typeof(PeakFormPatch).GetMethod(nameof(Postfix), BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
 
     public override IEnumerable<MethodBase> GetMethodsChecked() {
       yield return TargetMethodInfo;
@@ -29,19 +29,24 @@ namespace CommunityPatch.Patches {
     private static readonly byte[][] Hashes = {
       new byte[] {
         // e1.1.0.224785
-        0x52, 0xFC, 0xCB, 0x7C, 0xF9, 0x90, 0xBB, 0xAD,
-        0xDE, 0xDB, 0x69, 0xE9, 0x9A, 0x41, 0x28, 0x04,
-        0x7C, 0xE9, 0xDE, 0xBB, 0xD5, 0xD0, 0xCA, 0x3B,
-        0x4B, 0xE3, 0x19, 0x42, 0x1E, 0x36, 0xC8, 0xB2
+        0x92, 0xC3, 0x48, 0x33, 0x3C, 0x1A, 0x39, 0x52,
+        0xE5, 0x8B, 0xE4, 0x3E, 0xBD, 0x86, 0xFC, 0xA7,
+        0x56, 0x27, 0xD9, 0x3A, 0xA5, 0x53, 0xC4, 0xDF,
+        0xE2, 0x7B, 0x97, 0xBC, 0xB1, 0xAE, 0x36, 0x5E
       }
     };
 
     public override void Reset()
-      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "xIL6vOgI");
+      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "fBgGbxaw");
 
     public override bool IsApplicable(Game game)
       // ReSharper disable once CompareOfFloatsByEqualityOperator
     {
+      if (_perk == null)
+        return false;
+      if (_perk.PrimaryBonus != 0f)
+        return false;
+
       var patchInfo = Harmony.GetPatchInfo(TargetMethodInfo);
       if (AlreadyPatchedByOthers(patchInfo))
         return false;
@@ -51,12 +56,15 @@ namespace CommunityPatch.Patches {
     }
 
     public override void Apply(Game game) {
+      // Dear TaleWorlds; Value should probably be publicly exposed, maybe by a method
+      // and maybe marked [Obsolete] if you want to avoid your developers doing dirty deeds
       var textObjStrings = TextObject.ConvertToStringList(
         new List<TextObject> {
           _perk.Name,
           _perk.Description
         }
       );
+
       // most of the properties of skills have private setters, yet Initialize is public
       _perk.Initialize(
         textObjStrings[0],
@@ -64,29 +72,31 @@ namespace CommunityPatch.Patches {
         _perk.Skill,
         (int) _perk.RequiredSkillValue,
         _perk.AlternativePerk,
-        _perk.PrimaryRole, 0.1f,
+        _perk.PrimaryRole, 10f,
         _perk.SecondaryRole, _perk.SecondaryBonus,
-        SkillEffect.EffectIncrementType.AddFactor
+        _perk.IncrementType
       );
+
       if (Applied) return;
 
       CommunityPatchSubModule.Harmony.Patch(TargetMethodInfo,
         postfix: new HarmonyMethod(PatchMethodInfo));
+
       Applied = true;
     }
 
     // ReSharper disable once InconsistentNaming
     [MethodImpl(MethodImplOptions.NoInlining)]
-    private static void Postfix(ref int __result, Town town, StatExplainer explanation) {
-      var perk = ActivePatch._perk;
-      if (!(town?.Governor?.GetPerkValue(perk) ?? false))
-        return;
+    public static void Postfix(ref int __result, CharacterObject character, StatExplainer explanation) {
+      var result = __result;
 
-      var explainedNumber = new ExplainedNumber(__result, explanation);
-      if (explainedNumber.BaseNumber > 0) {
-        explainedNumber.AddFactor(perk.PrimaryBonus, perk.Name);
-        __result = (int) explainedNumber.ResultNumber;
-      }
+      var explainedNumber = new ExplainedNumber(result, explanation, null);
+
+      var perk = ActivePatch._perk;
+
+      PerkHelper.AddPerkBonusForCharacter(perk, character, ref explainedNumber);
+
+      __result = MBMath.Round(explainedNumber.ResultNumber);
     }
 
   }
