+++ src/CommunityPatch/Patches/Perks/Intelligence/Steward/AgrarianPatch.cs
@@ -1,58 +1,55 @@
-using System;
 using System.Collections.Generic;
 using System.Reflection;
-using System.Linq;
+using System.Runtime.CompilerServices;
 using TaleWorlds.CampaignSystem;
 using TaleWorlds.Core;
 using TaleWorlds.Localization;
 using HarmonyLib;
-using Helpers;
-using TaleWorlds.CampaignSystem.SandBox.GameComponents;
 using TaleWorlds.CampaignSystem.SandBox.GameComponents.Map;
-using TaleWorlds.Library;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Intelligence.Steward {
 
-  internal class StewardEnhancedMinesPatch : PatchBase<StewardEnhancedMinesPatch> {
+  public sealed class AgrarianPatch : PatchBase<AgrarianPatch> {
 
     public override bool Applied { get; protected set; }
 
     private static readonly MethodInfo TargetMethodInfo =
-      Type.GetType("TaleWorlds.CampaignSystem.SandBox.GameComponents.DefaultSettlementTaxModel, TaleWorlds.CampaignSystem")?
-        .GetMethod("CalculateVillageTaxFromIncome", BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
+      typeof(DefaultVillageProductionCalculatorModel).GetMethod(nameof(DefaultVillageProductionCalculatorModel.CalculateDailyFoodProductionAmount), BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(StewardEnhancedMinesPatch).GetMethod("Postfix", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
-    
+    private static readonly MethodInfo PatchMethodInfo = typeof(AgrarianPatch).GetMethod("Postfix", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+
     public override IEnumerable<MethodBase> GetMethodsChecked() {
       yield return TargetMethodInfo;
     }
-    
+
     private PerkObject _perk;
 
     private static readonly byte[][] Hashes = {
       new byte[] {
         // e1.1.0.224785
-        0x83, 0xCF, 0x66, 0x4A, 0x31, 0x3B, 0xF7, 0x98,
-        0xBE, 0xB8, 0x98, 0xA4, 0x85, 0x4D, 0xED, 0xA2,
-        0xCE, 0x37, 0xC1, 0x0E, 0x34, 0x2C, 0xB8, 0x84,
-        0x0E, 0xB2, 0x61, 0xA8, 0xB5, 0x97, 0x93, 0x08
+        0xB6, 0x97, 0x7F, 0x48, 0x9D, 0x7D, 0x8D, 0x7F,
+        0x17, 0x9B, 0x01, 0xB8, 0xC3, 0x6D, 0x87, 0x1B,
+        0x9E, 0xDC, 0xF4, 0xDA, 0x8B, 0xC2, 0xFD, 0x5C,
+        0xEE, 0x7C, 0x51, 0x93, 0x1D, 0x4E, 0x93, 0x5F
       }
     };
 
     public override void Reset()
-      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "6oE7rB6q");
+      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "XNc2NIGL");
 
     public override bool IsApplicable(Game game)
       // ReSharper disable once CompareOfFloatsByEqualityOperator
     {
       if (_perk == null)
         return false;
+      if (_perk.PrimaryBonus != 0f)
+        return false;
 
       var patchInfo = Harmony.GetPatchInfo(TargetMethodInfo);
       if (AlreadyPatchedByOthers(patchInfo))
         return false;
-      
+
       var hash = TargetMethodInfo.MakeCilSignatureSha256();
       return hash.MatchesAnySha256(Hashes);
     }
@@ -71,7 +68,7 @@ namespace CommunityPatch.Patches {
         _perk.Skill,
         (int) _perk.RequiredSkillValue,
         _perk.AlternativePerk,
-        _perk.PrimaryRole, 0.5f,
+        _perk.PrimaryRole, 0.3f,
         _perk.SecondaryRole, _perk.SecondaryBonus,
         SkillEffect.EffectIncrementType.AddFactor
       );
@@ -82,17 +79,15 @@ namespace CommunityPatch.Patches {
       Applied = true;
     }
 
-    // ReSharper disable once t
-    public static void Postfix(ref int __result, Village village, int marketIncome) {
+    // ReSharper disable once InconsistentNaming
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public static void Postfix(ref float __result, Village village) {
       var perk = ActivePatch._perk;
-      if (!(village.Bound?.OwnerClan?.Leader?.GetPerkValue(perk) ?? false)) {
+      if (!(village.Bound?.Town?.Governor?.GetPerkValue(perk) ?? false)) {
         return;
       }
-      if (village.VillageType.PrimaryProduction.IsFood) 
-        return;
-      var explainedNumber = new ExplainedNumber(__result, null);
-      explainedNumber.AddFactor(perk.PrimaryBonus,perk.Description);
-      __result = (int) explainedNumber.ResultNumber;
+
+      __result += (int) (__result * perk.PrimaryBonus);
     }
 
   }
