+++ src/CommunityPatch/Patches/Perks/Intelligence/Steward/BannerlordPerkPatch.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
@@ -9,15 +8,15 @@ using TaleWorlds.CampaignSystem.SandBox.GameComponents.Party;
 using TaleWorlds.Core;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Intelligence.Steward {
 
-  sealed class StewardSwordsAsTributePatch : PatchBase<StewardSwordsAsTributePatch> {
+  public sealed class BannerlordPatch : PatchBase<BannerlordPatch> {
 
     public override bool Applied { get; protected set; }
 
     private static readonly MethodInfo TargetMethodInfo = typeof(DefaultPartySizeLimitModel).GetMethod("CalculateMobilePartyMemberSizeLimit", BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly);
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(StewardSwordsAsTributePatch).GetMethod(nameof(Postfix), BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo PatchMethodInfo = typeof(BannerlordPatch).GetMethod(nameof(Postfix), BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
 
     public override IEnumerable<MethodBase> GetMethodsChecked() {
       yield return TargetMethodInfo;
@@ -34,7 +33,7 @@ namespace CommunityPatch.Patches {
         0x71, 0xD6, 0x8C, 0x75, 0xC4, 0xDF, 0x52, 0x5F
       },
       new byte[] {
-        // e1.0.6
+        // e1.0.5
         0x5A, 0xFE, 0xC3, 0xB1, 0x6A, 0xFC, 0xE0, 0xE0,
         0x43, 0x83, 0x5B, 0x73, 0x2F, 0x7D, 0x29, 0x9F,
         0x63, 0xAE, 0xD2, 0xC1, 0x6B, 0xE0, 0x0F, 0x32,
@@ -43,7 +42,7 @@ namespace CommunityPatch.Patches {
     };
 
     public override void Reset()
-      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "7fHHThQr");
+      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "MMv0U5Yr");
 
     public override void Apply(Game game) {
       if (Applied) return;
@@ -66,25 +65,15 @@ namespace CommunityPatch.Patches {
     [MethodImpl(MethodImplOptions.NoInlining)]
     private static void Postfix(ref int __result, MobileParty party, StatExplainer explanation) {
       var perk = ActivePatch._perk;
-      var hero = party.LeaderHero;
-      if (hero == null)
-        return;
-      if (!(hero?.GetPerkValue(perk) ?? false))
+      if (!(party.LeaderHero?.GetPerkValue(perk) ?? false))
         return;
 
-      var kingdomClans = hero.Clan?.Kingdom?.Clans;
-
-      if (kingdomClans == null)
-        return;
-
-      var extra = (int) Math.Max(0, (kingdomClans.Count() - 1) * perk.PrimaryBonus);
-
-      if (extra <= 0)
-        return;
-
-      var explainedNumber = new ExplainedNumber(__result, explanation);
-      explainedNumber.Add(extra, perk.Name);
-      __result = (int) explainedNumber.ResultNumber;
+      var extra = party.LeaderHero.Clan.Settlements.Count() * perk.PrimaryBonus;
+      if (extra > 0) {
+        var explainedNumber = new ExplainedNumber(__result, explanation);
+        explainedNumber.Add(party.LeaderHero.Clan.Settlements.Count() * perk.PrimaryBonus, perk.Name);
+        __result = (int) explainedNumber.ResultNumber;
+      }
     }
 
   }
