+++ src/CommunityPatch/Patches/Perks/Intelligence/Steward/EnhancedMinesPatch.cs
@@ -1,70 +1,69 @@
+using System;
 using System.Collections.Generic;
 using System.Reflection;
-using System.Runtime.CompilerServices;
+using System.Linq;
+using TaleWorlds.CampaignSystem;
+using TaleWorlds.Core;
+using TaleWorlds.Localization;
 using HarmonyLib;
 using Helpers;
-using TaleWorlds.CampaignSystem;
 using TaleWorlds.CampaignSystem.SandBox.GameComponents;
-using TaleWorlds.Core;
+using TaleWorlds.CampaignSystem.SandBox.GameComponents.Map;
 using TaleWorlds.Library;
-using TaleWorlds.Localization;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Intelligence.Steward {
 
-  internal class PeakFormPatch : PatchBase<PeakFormPatch> {
+  public sealed class EnhancedMinesPatch : PatchBase<EnhancedMinesPatch> {
 
     public override bool Applied { get; protected set; }
 
-    private static readonly MethodInfo TargetMethodInfo = typeof(DefaultCharacterStatsModel).GetMethod(nameof(DefaultCharacterStatsModel.MaxHitpoints), BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo TargetMethodInfo =
+      Type.GetType("TaleWorlds.CampaignSystem.SandBox.GameComponents.DefaultSettlementTaxModel, TaleWorlds.CampaignSystem")?
+        .GetMethod("CalculateVillageTaxFromIncome", BindingFlags.Public | BindingFlags.Instance | BindingFlags.DeclaredOnly);
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(PeakFormPatch).GetMethod(nameof(Postfix), BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
-
+    private static readonly MethodInfo PatchMethodInfo = typeof(EnhancedMinesPatch).GetMethod("Postfix", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+    
     public override IEnumerable<MethodBase> GetMethodsChecked() {
       yield return TargetMethodInfo;
     }
-
+    
     private PerkObject _perk;
 
     private static readonly byte[][] Hashes = {
       new byte[] {
         // e1.1.0.224785
-        0x92, 0xC3, 0x48, 0x33, 0x3C, 0x1A, 0x39, 0x52,
-        0xE5, 0x8B, 0xE4, 0x3E, 0xBD, 0x86, 0xFC, 0xA7,
-        0x56, 0x27, 0xD9, 0x3A, 0xA5, 0x53, 0xC4, 0xDF,
-        0xE2, 0x7B, 0x97, 0xBC, 0xB1, 0xAE, 0x36, 0x5E
+        0x83, 0xCF, 0x66, 0x4A, 0x31, 0x3B, 0xF7, 0x98,
+        0xBE, 0xB8, 0x98, 0xA4, 0x85, 0x4D, 0xED, 0xA2,
+        0xCE, 0x37, 0xC1, 0x0E, 0x34, 0x2C, 0xB8, 0x84,
+        0x0E, 0xB2, 0x61, 0xA8, 0xB5, 0x97, 0x93, 0x08
       }
     };
 
     public override void Reset()
-      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "fBgGbxaw");
+      => _perk = PerkObject.FindFirst(x => x.Name.GetID() == "6oE7rB6q");
 
     public override bool IsApplicable(Game game)
       // ReSharper disable once CompareOfFloatsByEqualityOperator
     {
       if (_perk == null)
         return false;
-      if (_perk.PrimaryBonus != 0f)
-        return false;
 
       var patchInfo = Harmony.GetPatchInfo(TargetMethodInfo);
       if (AlreadyPatchedByOthers(patchInfo))
         return false;
-
+      
       var hash = TargetMethodInfo.MakeCilSignatureSha256();
       return hash.MatchesAnySha256(Hashes);
     }
 
     public override void Apply(Game game) {
-      // Dear TaleWorlds; Value should probably be publicly exposed, maybe by a method
-      // and maybe marked [Obsolete] if you want to avoid your developers doing dirty deeds
       var textObjStrings = TextObject.ConvertToStringList(
         new List<TextObject> {
           _perk.Name,
           _perk.Description
         }
       );
-
       // most of the properties of skills have private setters, yet Initialize is public
       _perk.Initialize(
         textObjStrings[0],
@@ -72,31 +71,28 @@ namespace CommunityPatch.Patches {
         _perk.Skill,
         (int) _perk.RequiredSkillValue,
         _perk.AlternativePerk,
-        _perk.PrimaryRole, 10f,
+        _perk.PrimaryRole, 0.5f,
         _perk.SecondaryRole, _perk.SecondaryBonus,
-        _perk.IncrementType
+        SkillEffect.EffectIncrementType.AddFactor
       );
-
       if (Applied) return;
 
       CommunityPatchSubModule.Harmony.Patch(TargetMethodInfo,
         postfix: new HarmonyMethod(PatchMethodInfo));
-
       Applied = true;
     }
 
-    // ReSharper disable once InconsistentNaming
-    [MethodImpl(MethodImplOptions.NoInlining)]
-    public static void Postfix(ref int __result, CharacterObject character, StatExplainer explanation) {
-      var result = __result;
-
-      var explainedNumber = new ExplainedNumber(result, explanation, null);
-
+    // ReSharper disable once t
+    public static void Postfix(ref int __result, Village village, int marketIncome) {
       var perk = ActivePatch._perk;
-
-      PerkHelper.AddPerkBonusForCharacter(perk, character, ref explainedNumber);
-
-      __result = MBMath.Round(explainedNumber.ResultNumber);
+      if (!(village.Bound?.OwnerClan?.Leader?.GetPerkValue(perk) ?? false)) {
+        return;
+      }
+      if (village.VillageType.PrimaryProduction.IsFood) 
+        return;
+      var explainedNumber = new ExplainedNumber(__result, null);
+      explainedNumber.AddFactor(perk.PrimaryBonus,perk.Description);
+      __result = (int) explainedNumber.ResultNumber;
     }
 
   }
