+++ src/CommunityPatch/Patches/Perks/Intelligence/Steward/LogisticsExpertPatch.cs
@@ -7,9 +7,9 @@ using TaleWorlds.CampaignSystem.SandBox.GameComponents.Map;
 using TaleWorlds.Core;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Intelligence.Steward {
 
-  sealed class LogisticsExpertPatch : PatchBase<LogisticsExpertPatch> {
+  public sealed class LogisticsExpertPatch : PatchBase<LogisticsExpertPatch> {
 
     public override bool Applied { get; protected set; }
