+++ src/CommunityPatch/Patches/Perks/Intelligence/Steward/StewardTaxCollector.cs
@@ -7,9 +7,9 @@ using TaleWorlds.Core;
 using HarmonyLib;
 using static CommunityPatch.HarmonyHelpers;
 
-namespace CommunityPatch.Patches {
+namespace CommunityPatch.Patches.Perks.Intelligence.Steward {
 
-  sealed class StewardTaxCollectorPatch : PatchBase<StewardTaxCollectorPatch> {
+  public sealed class TaxCollectorPatch : PatchBase<TaxCollectorPatch> {
 
     public override bool Applied { get; protected set; }
 
@@ -17,7 +17,7 @@ namespace CommunityPatch.Patches {
       Type.GetType("TaleWorlds.CampaignSystem.SandBox.GameComponents.DefaultSettlementTaxModel, TaleWorlds.CampaignSystem")?
         .GetMethod("CalculateDailyTaxInternal", BindingFlags.NonPublic | BindingFlags.Instance | BindingFlags.DeclaredOnly);
 
-    private static readonly MethodInfo PatchMethodInfo = typeof(StewardTaxCollectorPatch).GetMethod(nameof(Postfix), BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
+    private static readonly MethodInfo PatchMethodInfo = typeof(TaxCollectorPatch).GetMethod(nameof(Postfix), BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static | BindingFlags.DeclaredOnly);
 
     public override IEnumerable<MethodBase> GetMethodsChecked() {
       yield return TargetMethodInfo;
